requirements.txt
altgraph==0.17.4
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.6
django-heroku==0.3.1
djangorestframework==3.14.0
gunicorn==21.2.0
image==1.5.33
packaging==23.2
pefile==2023.2.7
Pillow==10.0.1
psycopg2==2.9.9
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pytz==2023.3.post1
pywin32-ctypes==0.2.2
six==1.16.0
speedtest-cli==2.1.3
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.6.0
